remove `for_expr` submodule
